+++ CHANGELOG.md
@@ -19,6 +19,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Support for the `$signed` and `$unsigned` builtin functions.
 - Support auto-connected and unconnected ports.
 - Support for the `{...}` concatenation and `{N{...}}` repetition operators.
+- Add `-On` switch to control optimization level.
 
 ### Changed
 - Update llhd to v0.8.0.